refactor: add ArtworkDownloadButton
